[REGALLOC] making regalloc implementation base

+++ src/ir/ty.rs
@@ -1,6 +1,6 @@
 /// Saves the type name
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum TypeMetadata {
-    /// Signed 32 bit intenger
-    Int32,
+    /// Signed 64 bit intenger
+    Int64,
 }
